Split autodiffxd_test.cc into multiple translation units
