+++ drake/common/test/autodiffxd_addition_test.cc
@@ -1,0 +1,17 @@
+#include "drake/common/autodiff.h"
+#include "drake/common/test/autodiffxd_test.h"
+
+namespace drake {
+namespace test {
+namespace {
+
+TEST_F(AutoDiffXdTest, Addition) {
+  CHECK_BINARY_OP(+, x, y, 1.0);
+  CHECK_BINARY_OP(+, x, y, -1.0);
+  CHECK_BINARY_OP(+, y, x, 1.0);
+  CHECK_BINARY_OP(+, y, x, -1.0);
+}
+
+}  // namespace
+}  // namespace test
+}  // namespace drake
